+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- destroy: option -f can be used to delete/destroy a fscomp (no recursive support yet)
+
+### Changed
+- destroy: option -f (force) is now -F
 
 ## [0.5.9] 2018-12-16
 ### Added